requirements.txt
Flask 
bcrypt
Flask-Bcrypt 
Flask-Login 
Flask-Mail 
Flask-SQLAlchemy 
Flask-WTF 
WTForms
pymysql
Pillow
email_validator
requests